Rework format-string handling, part 1.
- Support multi-character variable names (e.g. "%{foo}").
- Support conditionals (e.g. "%{?foo,yes,no}").
- Support the zero flag to specify zero padding (e.g. "%02f").
- As a result, a variable width is now indicated by an asterisk rather than a
  zero (e.g. "%*f" rather than "%0f").

Because the format-string syntax has become more complex and format strings
will be used more often, a format string first is parsed once and broken into a
linked list of field parts and literal parts. Then, every time an output string
needs to be constructed from a specified set of format variables, the linked
list is traversed. Literal parts are inserted into the output string verbatim.
For field parts the value is looked up in the specified set of format variables
and inserted into the output string.