Fix playback of samples of 8 bits or less on ao and sndio.
ao and sndio expect 1 byte per sample if it has a bit depth of 8 or less.
However, Siren currently always stores a sample in 2 bytes. Therefore, for ao,
always specify a bit depth of 16, regardless of the true bit depth. For sndio,
don't change the bit depth, but always specify 2 bytes per sample.