Fix playback of certain WavPack files on sndio.
WavPack aligns samples whose bit depth is not a multiple of 8 on the MSB rather
than the LSB. For example, a 12-bit sample would be aligned on bit 15 rather
than bit 0. This confuses sndio, which by default expects such a sample to be
aligned on the LSB.

Therefore, determine the bit depth from the number of bytes per sample. Thus, a
12-bit sample, for example, would be treated as if it were a 16-bit sample.