Don't allow the oss plug-in to change the volume level while the device is closed.
Currently, the oss plug-in saves the current volume level if the device is
closed and restores it when the device is opened again. This allows the user to
change the volume level while the device is closed: in that case only the saved
volume level is changed.

However, the volume level can also be changed externally (e.g. by a mixer
utility) and unfortunately the OSS 4 API does not offer a notification
mechanism for these changes. Thus, when the oss plug-in restores the saved
volume level, it possibly ignores and undoes a volume-level change made
externally. This will probably cause surprise and annoyance.

Therefore, no longer allow the volume level to be changed while the device is
closed so that it isn't necessary to set the volume level when the device is
opened.