Rework the way in which tracks and metadata are handled.
Until now, cache.c contained a tree with metadata entries. On start-up, this
tree was populated with the entries from the metadata-cache file. Whenever a
track was requested with track_init(), memory was allocated for the track and
its metadata copied from the tree to the track.

Now, the tree has been moved from cache.c to track.c and contains not only
metadata, but full tracks. On start-up, this tree is populated with tracks
based on the entries from the metadata-cache file. Whenever a track is
requested with track_get() (which replaces track_init()), the track is looked
up in the tree and a pointer to the track is returned.

This change allows for the following two improvements.

- It will possible to update the metadata, because they are referenced by
  tracks rather than copied to.
- Memory usage has significantly decreased, because metadata are no longer
  copied and memory for a track is allocated only once.