Prevent a potential mutex deadlock.
The deadlock occurs when screen_prompt_end() locks screen_curses_mtx and then
attempts to lock option_tree_mtx while at the same time player_print_status()
or player_print_track() attempts to lock them in reverse order.

The deadlock is reproducible by repeatedly opening and closing a prompt while
playing back a track until the timing is right. The best chance of success is
with Linux and the libao null driver.

It seems the best solution is to ensure the screen.c functions do not attempt
to hold these mutexes simultaneously.