Tracks should not silently disappear from the library if they cannot be found on the file system or if no input plug-in for them can be found.
Therefore, introduce track_require() which (unlike track_get()) returns a track
even if it does not exist or if no input plug-in for it can be found, and use
it to load the library.

Additionally, with this change it no longer is necessary always to save the
library upon exit. It now is saved only if it has been modified.