Essentially revert commit d7bc6798ffdb.
OSS on FreeBSD resets the volume when the device is opened. Since the oss
plug-in closes the device after playback has finished (so that other programs
can use it), the volume is reset whenever a new track is played back. This is
very annoying, so ensure the original volume is set when the device is opened.

The drawback is that volume changes by other programs (e.g. mixer programs) may
be lost, but that is far less annoying.