Rework how the ffmpeg plug-in is used.
The ffmpeg plug-in is currently used as a fallback plug-in: it is given every
file not supported by another plug-in. This has undesirable consequences. For
example, the add-path command adds non-audio files to the library.

So drop the idea of a fallback plug-in and use the ffmpeg plug-in like all
others. This means that the ffmpeg plug-in now specifies a list of supported
file extensions.

Using file extensions to associate files with plug-ins is not ideal, but it
works fine most of the time and is less expensive than file probing.